Remove network from AbstractBaseVoltageDiagramStyleProvider
